requirements.txt
asgiref==3.7.2
Django==4.2.4
django-environ==0.11.2
django-extensions==3.2.3
django-rest-framework==0.1.0
djangorestframework==3.14.0
mysqlclient==2.2.0
pytz==2023.3
sqlparse==0.4.4
typing_extensions==4.7.1